requirements.txt
numpy==1.21.6
onnxruntime==1.13.1
opencv_python==4.6.0.66
openvino==2022.2.0
Pillow==9.3.0
PyQt5==5.15.7
pyserial==3.5
torch==1.8.1
torchvision==0.9.1
